lifting up the function find_offloadable_tree(iet) to ir/support/
